+++ ChangeLog.md
@@ -3,6 +3,12 @@ REST Client and Server APIs for the XP Framework ChangeLog
 
 ## ?.?.? / ????-??-??
 
+## 8.1.0 / 2016-09-24
+
+* Changed `RestRequest` to optionally accept parameters in constructor
+  argument uri, adding support to the `resource()` method by doing so.
+  (@thekid)
+
 ## 8.0.0 / 2016-09-04
 
 * Merged PR #16: Resource API. Entry point is `webservices.rest.Endpoint`.